requirements.txt
Django==5.2.4
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
Pillow==11.1.0
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==23.0.0